% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromort.R
\name{as_probability}
\alias{as_probability}
\title{Convert Micromorts to Probability}
\usage{
as_probability(micromorts)
}
\arguments{
\item{micromorts}{Numeric. Risk in micromorts.}
}
\value{
Numeric probability.
}
\description{
Convert Micromorts to Probability
}
\examples{
as_probability(1) # 1e-6
}
\seealso{
\code{\link[=as_micromort]{as_micromort()}}, \code{\link[=as_microlife]{as_microlife()}}

Other conversion: 
\code{\link{as_microlife}()},
\code{\link{as_micromort}()},
\code{\link{lle}()},
\code{\link{value_of_micromort}()}
}
\concept{conversion}
