% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lifestyle_tradeoff}
\alias{lifestyle_tradeoff}
\title{Calculate Lifestyle Tradeoff}
\usage{
lifestyle_tradeoff(bad_habit, good_habit)
}
\arguments{
\item{bad_habit}{Factor name of the bad habit (from \code{\link[=chronic_risks]{chronic_risks()}})}

\item{good_habit}{Factor name of the compensating behavior (from \code{\link[=chronic_risks]{chronic_risks()}})}
}
\value{
A tibble showing the tradeoff ratio
}
\description{
Calculate how much of one good habit compensates for a bad habit.
}
\examples{
# How much exercise offsets 2 cigarettes?
lifestyle_tradeoff("Smoking 2 cigarettes", "20 min moderate exercise")
}
\seealso{
\code{\link[=compare_interventions]{compare_interventions()}}, \code{\link[=chronic_risks]{chronic_risks()}}

Other analysis: 
\code{\link{annual_risk_budget}()},
\code{\link{compare_interventions}()},
\code{\link{daily_hazard_rate}()},
\code{\link{risk_sensitivity}()},
\code{\link{toxicological_risk}()}
}
\concept{analysis}
