% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromort.R
\name{lle}
\alias{lle}
\title{Loss of Life Expectancy (LLE)}
\usage{
lle(prob, life_expectancy = 40)
}
\arguments{
\item{prob}{Numeric. Probability of death.}

\item{life_expectancy}{Numeric. Remaining life expectancy in years (default 40).}
}
\value{
Numeric. Loss of life expectancy in seconds, minutes, or days (estimated).
}
\description{
Estimates the average time lost from a lifespan due to a specific risk.
}
\examples{
lle(1/1000000, 40) # Loss from 1 micromort
}
