% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromort.R
\name{as_microlife}
\alias{as_microlife}
\title{Convert to Microlives}
\usage{
as_microlife(minutes_lost)
}
\arguments{
\item{minutes_lost}{Numeric. Life expectancy lost in minutes.}
}
\value{
Numeric. Value in microlives.
}
\description{
A microlife represents a 30-minute change in life expectancy.
This function estimates the impact of a chronic risk in microlives.
}
\examples{
as_microlife(30) # 1 microlife
}
