% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromort.R
\name{as_micromort}
\alias{as_micromort}
\title{Convert Probability to Micromorts}
\usage{
as_micromort(prob)
}
\arguments{
\item{prob}{Numeric. Probability of death (0 to 1).}
}
\value{
Numeric value in micromorts.
}
\description{
A micromort represents a one-in-a-million chance of death.
This function converts a raw probability of death into micromorts.
}
\examples{
as_micromort(1/1000000) # 1 micromort
as_micromort(1/10000)   # 100 micromorts
}
