% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromort.R
\name{as_probability}
\alias{as_probability}
\title{Convert Micromorts to Probability}
\usage{
as_probability(micromorts)
}
\arguments{
\item{micromorts}{Numeric. Risk in micromorts.}
}
\value{
Numeric probability.
}
\description{
Convert Micromorts to Probability
}
\examples{
as_probability(1) # 1e-6
}
