% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risks.R
\name{common_risks}
\alias{common_risks}
\title{Common Risks in Micromorts}
\usage{
common_risks()
}
\value{
A tibble with columns: activity, risk_micromorts, category.
}
\description{
A dataset of common activities and their associated risk in micromorts.
Source: Wikipedia and other public risk assessments.
}
\examples{
common_risks()
}
