% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_risks}
\alias{plot_risks}
\title{Plot Risk Comparison}
\usage{
plot_risks(risks = common_risks(), facet = TRUE)
}
\arguments{
\item{risks}{Tibble. Dataframe of risks, defaults to common_risks().}

\item{facet}{Logical. If TRUE, splits plot into COVID-19 and Other panels (2x1 stacked).
Default is TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Visualizes the risk of different activities in micromorts.
}
\examples{
plot_risks()
plot_risks(facet = FALSE)
}
