% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_risks}
\alias{plot_risks}
\title{Plot Risk Comparison}
\usage{
plot_risks(risks = common_risks())
}
\arguments{
\item{risks}{Tibble. Dataframe of risks, defaults to common_risks().}
}
\value{
A ggplot2 object.
}
\description{
Visualizes the risk of different activities in micromorts.
}
\examples{
plot_risks()
}
