% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromort.R
\name{value_of_micromort}
\alias{value_of_micromort}
\title{Value of a Statistical Life (VSL) to Micromort Value}
\usage{
value_of_micromort(vsl = 1e+07)
}
\arguments{
\item{vsl}{Numeric. Value of a Statistical Life (default $10,000,000).}
}
\value{
Numeric value of one micromort.
}
\description{
Calculates the monetary value of one micromort based on the Value of a Statistical Life (VSL).
}
\examples{
value_of_micromort(10000000) # $10
}
